+++ tanks-game/constants.cpp
@@ -1,4 +1,5 @@
 #include "constants.h"
+
 #define _USE_MATH_DEFINES
 #include <cmath>
 
@@ -9,14 +10,14 @@ const double TANK_RADIUS = 16;
 const double SHOOT_COOLDOWN = 100;
 const double POWER_COOLDOWN = 500; //does nothing
 
-const double DESTRUCTION_TIER = 2;
+const float DESTRUCTION_TIER = 2;
 const double BULLET_TO_TANK_RADIUS_RATIO = .25;
 const double BULLET_TO_TANK_SPEED_RATIO = 4; //2 isn't that bad
 
-const double LOW_TIER = -999;
-const double HIGH_TIER = 999;
-const double LOW_IMPORTANCE = -1;
-const double HIGH_IMPORTANCE = 999;
+const float LOW_TIER = -999;
+const float HIGH_TIER = 999;
+const float LOW_IMPORTANCE = -1;
+const float HIGH_IMPORTANCE = 999;
 
 const Team_ID DEFAULT_TEAM = 0;
 const Team_ID HAZARD_TEAM = -1;
