+++ tanks-game/invincible-named-power.h
@@ -44,10 +44,10 @@ public:
 	virtual double getTankMaxSpeedMultiplier() const override { return .5; }
 	virtual double getTankAccelerationMultiplier() const override { return .5; }
 
-	virtual double getOffenseImportance() const override { return 0; }
-	virtual double getOffenseTier(const Tank*) const override { return 0; }
-	virtual double getDefenseImportance() const override { return 0; }
-	virtual double getDefenseTier(const Tank*) const override { return 2.5; } // this is not DESTRUCTION_TIER+.5 in case someone wants to have a higher destruction tier
+	virtual float getOffenseImportance() const override { return 0; }
+	virtual float getOffenseTier(const Tank*) const override { return 0; }
+	virtual float getDefenseImportance() const override { return 0; }
+	virtual float getDefenseTier(const Tank*) const override { return 2.5; } // this is not DESTRUCTION_TIER+.5 in case someone wants to have a higher destruction tier
 
 	InvincibleNamedTankPower();
 };
@@ -66,10 +66,10 @@ public:
 	virtual BulletPower* makeDuplicate() const override { return new InvincibleNamedBulletPower(); }
 	virtual TankPower* makeTankPower() const override;
 
-	virtual double getOffenseImportance() const override { return 0; }
-	virtual double getOffenseTier(const Bullet*) const override { return 1; }
-	virtual double getDefenseImportance() const override { return 0; }
-	virtual double getDefenseTier(const Bullet*) const override { return 1; }
+	virtual float getOffenseImportance() const override { return 0; }
+	virtual float getOffenseTier(const Bullet*) const override { return 1; }
+	virtual float getDefenseImportance() const override { return 0; }
+	virtual float getDefenseTier(const Bullet*) const override { return 1; }
 
 	InvincibleNamedBulletPower();
 };
