+++ tanks-game/mega-death-power.h
@@ -3,7 +3,7 @@
 
 class MegaDeathPower : public Power {
 public: //bullet stuff
-	static const double destroyWallTier; //(and destroyHazardTier)
+	static const float destroyWallTier; //(and destroyHazardTier)
 	static const double bulletSizeMultiplierPerTick;
 
 public:
@@ -80,10 +80,10 @@ public:
 	virtual double getBulletAcceleration() const override { return 1.0/64; }
 	//with bulletSizeMultiplierPerTick = 257.0/256.0, 1.0/512 or 1.0/1024 (with a smaller radius multiplier per tick) might be closer to what I orignially thought of
 
-	virtual double getOffenseImportance() const override { return 0; }
-	virtual double getOffenseTier(const Bullet*) const override;
-	virtual double getDefenseImportance() const override { return 0; }
-	virtual double getDefenseTier(const Bullet*) const override;
+	virtual float getOffenseImportance() const override { return 0; }
+	virtual float getOffenseTier(const Bullet*) const override;
+	virtual float getDefenseImportance() const override { return 0; }
+	virtual float getDefenseTier(const Bullet*) const override;
 
 	MegaDeathBulletPower();
 };
